Manager changes
Classic* are managers
Simple*Manager removed
Tests fixed
