+++ src/main/java/org/jpokemon/api/classic/ClassicTypes.java
@@ -1,39 +1,37 @@
 package org.jpokemon.api.classic;
 
+import java.lang.reflect.Field;
+import java.lang.reflect.Modifier;
+import java.util.HashMap;
+import java.util.Map;
+
+import org.jpokemon.api.JPokemonError;
+import org.jpokemon.api.Manager;
 import org.jpokemon.api.types.PokemonType;
 
 /**
- * Defines the list of the 17 classic Pokémon types. This class functions very
- * much like an enum, providing constants that can be referenced by Moves and
- * Species.
+ * Defines a "simplest-possible" implementation of the {@link Manager} interface
+ * for types. This manager provides the classic types. See {@link #init}
  * 
- * If you wish to add a new type, simply instantiate one and give it a name. To
- * modify a type, you can modify the static field of this class that refers to
- * it, or retrieve the instance by name with `getType`. Both are equivalent.
- * 
- * @author atheriel@gmail.com
- * @author Zach Taylor
- * 
- * @since  0.1
+ * Note that this class is final. If you want to add or remove natures, it
+ * cannot be extended; you must write your own lookup class.
  */
-public class ClassicTypes {
-	/** Provides the classic type `????`. */
+public final class ClassicTypes implements Manager<PokemonType> {
+	private final Map<String, PokemonType> typeList = new HashMap<String, PokemonType>();
+
 	public static final PokemonType TYPELESS = new TypelessType();
 
 	/** Provides the classic type `NORMAL`. */
 	public static final PokemonType NORMAL = new PokemonType().setName("Normal")
-			.setNotVeryEffectiveAgainst("Rock", "Steel")
-			.setIneffectiveAgainst("Ghost");
+			.setNotVeryEffectiveAgainst("Rock", "Steel").setIneffectiveAgainst("Ghost");
 
 	/** Provides the classic type `FIRE`. */
 	public static final PokemonType FIRE = new PokemonType().setName("Fire")
-			.setSuperEffectiveAgainst("Grass", "Bug", "Steel")
-			.setNotVeryEffectiveAgainst("Fire", "Water", "Rock", "Dragon");
+			.setSuperEffectiveAgainst("Grass", "Bug", "Steel").setNotVeryEffectiveAgainst("Fire", "Water", "Rock", "Dragon");
 
 	/** Provides the classic type `WATER`. */
 	public static final PokemonType WATER = new PokemonType().setName("Water")
-			.setSuperEffectiveAgainst("Fire", "Ground", "Rock")
-			.setNotVeryEffectiveAgainst("Water", "Grass", "Dragon");;
+			.setSuperEffectiveAgainst("Fire", "Ground", "Rock").setNotVeryEffectiveAgainst("Water", "Grass", "Dragon");;
 
 	/** Provides the classic type `GRASS`. */
 	public static final PokemonType GRASS = new PokemonType().setName("Grass")
@@ -42,8 +40,7 @@ public class ClassicTypes {
 
 	/** Provides the classic type `ELECTRIC`. */
 	public static final PokemonType ELECTRIC = new PokemonType().setName("Electric")
-			.setSuperEffectiveAgainst("Water", "Flying")
-			.setNotVeryEffectiveAgainst("Electric", "Grass", "Dragon")
+			.setSuperEffectiveAgainst("Water", "Flying").setNotVeryEffectiveAgainst("Electric", "Grass", "Dragon")
 			.setIneffectiveAgainst("Ground");
 
 	/** Provides the classic type `ICE`. */
@@ -54,30 +51,24 @@ public class ClassicTypes {
 	/** Provides the classic type `FIGHTING`. */
 	public static final PokemonType FIGHTING = new PokemonType().setName("Fighting")
 			.setSuperEffectiveAgainst("Normal", "Ice", "Rock", "Dark", "Steel")
-			.setNotVeryEffectiveAgainst("Poison", "Flying", "Psychic", "Bug")
-			.setIneffectiveAgainst("Ghost");
+			.setNotVeryEffectiveAgainst("Poison", "Flying", "Psychic", "Bug").setIneffectiveAgainst("Ghost");
 
 	/** Provides the classic type `POISON`. */
-	public static final PokemonType POISON = new PokemonType().setName("Poison")
-			.setSuperEffectiveAgainst("Grass")
-			.setNotVeryEffectiveAgainst("Poison", "Ground", "Rock", "Ghost")
-			.setIneffectiveAgainst("Steel");
+	public static final PokemonType POISON = new PokemonType().setName("Poison").setSuperEffectiveAgainst("Grass")
+			.setNotVeryEffectiveAgainst("Poison", "Ground", "Rock", "Ghost").setIneffectiveAgainst("Steel");
 
 	/** Provides the classic type `GROUND`. */
 	public static final PokemonType GROUND = new PokemonType().setName("Ground")
 			.setSuperEffectiveAgainst("Fire", "Electric", "Poison", "Rock", "Steel")
-			.setNotVeryEffectiveAgainst("Bug", "Grass")
-			.setIneffectiveAgainst("Flying");
+			.setNotVeryEffectiveAgainst("Bug", "Grass").setIneffectiveAgainst("Flying");
 
 	/** Provides the classic type `FLYING`. */
 	public static final PokemonType FLYING = new PokemonType().setName("Flying")
-			.setSuperEffectiveAgainst("Grass", "Fighting", "Bug")
-			.setNotVeryEffectiveAgainst("Electric", "Rock", "Steel");
+			.setSuperEffectiveAgainst("Grass", "Fighting", "Bug").setNotVeryEffectiveAgainst("Electric", "Rock", "Steel");
 
 	/** Provides the classic type `PSYCHIC`. */
 	public static final PokemonType PSYCHIC = new PokemonType().setName("Psychic")
-			.setSuperEffectiveAgainst("Fighting", "Poison")
-			.setNotVeryEffectiveAgainst("Psychic", "Steel")
+			.setSuperEffectiveAgainst("Fighting", "Poison").setNotVeryEffectiveAgainst("Psychic", "Steel")
 			.setIneffectiveAgainst("Dark");
 
 	/** Provides the classic type `BUG`. */
@@ -92,26 +83,75 @@ public class ClassicTypes {
 
 	/** Provides the classic type `GHOST`. */
 	public static final PokemonType GHOST = new PokemonType().setName("Ghost")
-			.setSuperEffectiveAgainst("Ghost", "Psychic")
-			.setNotVeryEffectiveAgainst("Dark", "Steel")
+			.setSuperEffectiveAgainst("Ghost", "Psychic").setNotVeryEffectiveAgainst("Dark", "Steel")
 			.setIneffectiveAgainst("Normal");
 
 	/** Provides the classic type `DRAGON`. */
-	public static final PokemonType DRAGON = new PokemonType().setName("Dragon")
-			.setSuperEffectiveAgainst("Dragon")
+	public static final PokemonType DRAGON = new PokemonType().setName("Dragon").setSuperEffectiveAgainst("Dragon")
 			.setNotVeryEffectiveAgainst("Steel");
 
 	/** Provides the classic type `DARK`. */
-	public static final PokemonType DARK = new PokemonType().setName("Dark")
-			.setSuperEffectiveAgainst("Psychic", "Ghost")
+	public static final PokemonType DARK = new PokemonType().setName("Dark").setSuperEffectiveAgainst("Psychic", "Ghost")
 			.setNotVeryEffectiveAgainst("Fighting", "Dark", "Steel");
 
 	/** Provides the classic type `STEEL`. */
-	public static final PokemonType STEEL = new PokemonType().setName("Steel")
-			.setSuperEffectiveAgainst("Ice", "Rock")
+	public static final PokemonType STEEL = new PokemonType().setName("Steel").setSuperEffectiveAgainst("Ice", "Rock")
 			.setNotVeryEffectiveAgainst("Fire", "Water", "Electric", "Steel");
 
-	/** Provides a private constructor. */
+	@Override
+	public boolean register(PokemonType type) throws JPokemonError {
+		if (typeList.containsKey(type.getName())) {
+			throw new JPokemonError("A type with the name " + type.getName() + " has already been registered!");
+		}
+		if (typeList.containsValue(type)) {
+			throw new JPokemonError("This type is already registered!");
+		}
+		typeList.put(type.getName(), type);
+		return true;
+	}
+
+	@Override
+	public boolean isRegistered(PokemonType type) {
+		return typeList.containsValue(type);
+	}
+
+	@Override
+	public PokemonType getByName(String name) {
+		if (!typeList.containsKey(name)) {
+			return null;
+		}
+		return typeList.get(name);
+	}
+
+	/**
+	 * Initializes {@link PokemonType#manager} to be an instance of ClassicTypes,
+	 * with all classic types populated in the manager.
+	 */
+	public static void init() {
+		if (PokemonType.manager != null) {
+			throw new JPokemonError("PokemonType.manager already defined.");
+		}
+
+		PokemonType.manager = new ClassicTypes();
+
+		for (Field field : ClassicTypes.class.getFields()) {
+			try {
+				if ((field.getModifiers() & (Modifier.PUBLIC | Modifier.STATIC | Modifier.FINAL)) <= 0) {
+					continue;
+				}
+
+				Object temp = field.get(null);
+
+				if (temp instanceof PokemonType) {
+					PokemonType type = (PokemonType) temp;
+					PokemonType.manager.register(type);
+				}
+			} catch (IllegalAccessException exception) {
+			}
+		}
+	}
+
+	/** Provides the private constructor. */
 	private ClassicTypes() {
 	}
 }
